+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "ts-node src/testindex.ts"
+    "start": "ts-node src/index.ts"
   },
   "author": "",
   "license": "ISC",
